feat(i18n): model Quran source messages
